+++ frontend/src/context/AuthContext.jsx
@@ -5,26 +5,28 @@ const AuthContext = createContext();
 
 // Provider component
 export const AuthProvider = ({ children }) => {
-    const [usertoken, setToken] = useState(() => localStorage.getItem("usertoken") || null);
+    const [token, setToken] = useState(() => localStorage.getItem("token") || null);
     const [user, setUser] = useState(() => JSON.parse(localStorage.getItem("user")) || null);
 
-    const login = (usertoken, userData) => {
-        setToken(usertoken);
+    const login = (token, userData) => {
+        setToken(token);
         setUser(userData);
-        localStorage.setItem("usertoken", usertoken);
+        localStorage.setItem("token", token);
         localStorage.setItem("user", JSON.stringify(userData));
         localStorage.setItem("userRole", userData.role);
     };
-
+    
     const logout = () => {
         setToken(null);
         setUser(null);
-        localStorage.removeItem("usertoken");
+        localStorage.removeItem("token");
         localStorage.removeItem("user");
+        localStorage.removeItem("userRole");
     };
+    
 
     return (
-        <AuthContext.Provider value={{ usertoken, user, login, logout }}>
+        <AuthContext.Provider value={{ token, user, login, logout }}>
             {children}
         </AuthContext.Provider>
     );